+++ src/edu/jhu/thrax/hadoop/features/mapred/LhsGivenTargetPhraseFeature.java
@@ -5,7 +5,6 @@ import java.util.Arrays;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.io.FloatWritable;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Writable;
 import org.apache.hadoop.io.WritableComparator;
 import org.apache.hadoop.io.WritableUtils;
@@ -42,15 +41,15 @@ public class LhsGivenTargetPhraseFeature extends MapReduceFeature {
     return RuleWritable.TargetPartitioner.class;
   }
 
-  public Class<? extends Mapper<RuleWritable, Annotation, RuleWritable, IntWritable>> mapperClass() {
+  public Class<? extends Mapper<RuleWritable, Annotation, RuleWritable, FloatWritable>> mapperClass() {
     return Map.class;
   }
 
-  public Class<? extends Reducer<RuleWritable, IntWritable, RuleWritable, FeaturePair>> reducerClass() {
+  public Class<? extends Reducer<RuleWritable, FloatWritable, RuleWritable, FeaturePair>> reducerClass() {
     return Reduce.class;
   }
 
-  private static class Map extends Mapper<RuleWritable, Annotation, RuleWritable, IntWritable> {
+  private static class Map extends Mapper<RuleWritable, Annotation, RuleWritable, FloatWritable> {
 
     protected void map(RuleWritable key, Annotation value, Context context) throws IOException,
         InterruptedException {
@@ -64,7 +63,7 @@ public class LhsGivenTargetPhraseFeature extends MapReduceFeature {
       lhs_target_marginal.source = PrimitiveArrayMarginalComparator.MARGINAL;
       target_marginal.monotone = false;
 
-      IntWritable count = new IntWritable(value.count());
+      FloatWritable count = new FloatWritable(value.count());
 
       context.write(key, count);
       context.write(lhs_target_marginal, count);
@@ -72,7 +71,7 @@ public class LhsGivenTargetPhraseFeature extends MapReduceFeature {
     }
   }
 
-  private static class Reduce extends Reducer<RuleWritable, IntWritable, RuleWritable, FeaturePair> {
+  private static class Reduce extends Reducer<RuleWritable, FloatWritable, RuleWritable, FeaturePair> {
 
     protected void setup(Context context) throws IOException, InterruptedException {
       Configuration conf = context.getConfiguration();
@@ -80,15 +79,15 @@ public class LhsGivenTargetPhraseFeature extends MapReduceFeature {
       Vocabulary.initialize(conf, vocabulary_path);
     }
 
-    private int marginal;
+    private float marginal;
     private FloatWritable prob;
 
-    protected void reduce(RuleWritable key, Iterable<IntWritable> values, Context context)
+    protected void reduce(RuleWritable key, Iterable<FloatWritable> values, Context context)
         throws IOException, InterruptedException {
       if (key.lhs == PrimitiveUtils.MARGINAL_ID) {
         // We only get here if it is the very first time we saw the source.
         marginal = 0;
-        for (IntWritable x : values)
+        for (FloatWritable x : values)
           marginal += x.get();
         return;
       }
@@ -97,9 +96,9 @@ public class LhsGivenTargetPhraseFeature extends MapReduceFeature {
       if (Arrays.equals(key.source, PrimitiveArrayMarginalComparator.MARGINAL)) {
         // We only get in here if it's a new LHS.
         int count = 0;
-        for (IntWritable x : values)
+        for (FloatWritable x : values)
           count += x.get();
-        prob = new FloatWritable((float) -Math.log(count / (float) marginal));
+        prob = new FloatWritable((float) -Math.log(count / marginal));
         return;
       }
       context.write(key, new FeaturePair(Vocabulary.id(LABEL), prob));
