+++ src/edu/jhu/thrax/hadoop/features/mapred/SourcePhraseGivenTargetFeature.java
@@ -5,7 +5,6 @@ import java.util.Arrays;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.io.FloatWritable;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Writable;
 import org.apache.hadoop.io.WritableComparator;
 import org.apache.hadoop.io.WritableUtils;
@@ -51,7 +50,7 @@ public class SourcePhraseGivenTargetFeature extends MapReduceFeature {
     return Reduce.class;
   }
 
-  private static class Map extends Mapper<RuleWritable, Annotation, RuleWritable, IntWritable> {
+  private static class Map extends Mapper<RuleWritable, Annotation, RuleWritable, FloatWritable> {
 
     protected void setup(Context context) throws IOException, InterruptedException {
       Configuration conf = context.getConfiguration();
@@ -69,7 +68,7 @@ public class SourcePhraseGivenTargetFeature extends MapReduceFeature {
       RuleWritable lhs_marginal = new RuleWritable(key);
       lhs_marginal.lhs = PrimitiveUtils.MARGINAL_ID;
 
-      IntWritable count = new IntWritable(value.count());
+      FloatWritable count = new FloatWritable(value.count());
 
       context.write(target_marginal, count);
       context.write(lhs_marginal, count);
@@ -77,8 +76,8 @@ public class SourcePhraseGivenTargetFeature extends MapReduceFeature {
     }
   }
 
-  private static class Reduce extends Reducer<RuleWritable, IntWritable, RuleWritable, FeaturePair> {
-    private int marginal;
+  private static class Reduce extends Reducer<RuleWritable, FloatWritable, RuleWritable, FeaturePair> {
+    private float marginal;
     private FloatWritable prob;
 
     protected void setup(Context context) throws IOException, InterruptedException {
@@ -87,19 +86,19 @@ public class SourcePhraseGivenTargetFeature extends MapReduceFeature {
       Vocabulary.initialize(conf, vocabulary_path);
     }
 
-    protected void reduce(RuleWritable key, Iterable<IntWritable> values, Context context)
+    protected void reduce(RuleWritable key, Iterable<FloatWritable> values, Context context)
         throws IOException, InterruptedException {
       if (Arrays.equals(key.source, PrimitiveArrayMarginalComparator.MARGINAL)) {
         marginal = 0;
-        for (IntWritable x : values)
+        for (FloatWritable x : values)
           marginal += x.get();
         return;
       }
       if (key.lhs == PrimitiveUtils.MARGINAL_ID) {
-        int count = 0;
-        for (IntWritable x : values)
+        float count = 0;
+        for (FloatWritable x : values)
           count += x.get();
-        prob = new FloatWritable((float) -Math.log(count / (float) marginal));
+        prob = new FloatWritable((float) -Math.log(count / marginal));
         return;
       }
       context.write(key, new FeaturePair(Vocabulary.id(LABEL), prob));
